+++ cg/meta/demultiplex/demux_post_processing.py
@@ -1,4 +1,5 @@
 import logging
+import shutil
 from pathlib import Path
 from typing import Iterable, List, Optional
 
@@ -114,6 +115,25 @@ class DemuxPostProcessingAPI:
             report_path: Path = demux_results.demux_dir / f"stats-{project_name}-{flowcell_id}.txt"
             self.write_report(report_path=report_path, report_data=report_data)
 
+    @staticmethod
+    def copy_sample_sheet(demux_results: DemuxResults) -> None:
+        """Copy the sample sheet from run dir to demux dir"""
+        LOG.info(
+            "Copy sample sheet (%s) from flowcell to demuxed result dir (%s)",
+            demux_results.sample_sheet_path,
+            demux_results.demux_sample_sheet_path,
+        )
+        shutil.copy(
+            demux_results.sample_sheet_path.as_posix(),
+            demux_results.demux_sample_sheet_path.as_posix(),
+        )
+
+    @staticmethod
+    def create_copy_complete_file(demux_results: DemuxResults) -> None:
+        """Suggestion is to deprecate this as a flag and store information in database instead"""
+        LOG.info("Creating %s", demux_results.copy_complete_path)
+        demux_results.copy_complete_path.touch()
+
     def post_process_flowcell(self, demux_results: DemuxResults) -> None:
         """Run all the necessary steps for post processing a demultiplexed flowcell
 
@@ -128,6 +148,8 @@ class DemuxPostProcessingAPI:
             self.rename_files(demux_results=demux_results)
         self.add_to_cgstats(demux_results=demux_results)
         self.create_cgstats_reports(demux_results=demux_results)
+        self.copy_sample_sheet(demux_results=demux_results)
+        self.create_copy_complete_file(demux_results=demux_results)
 
     def finish_flowcell(self, flowcell_name: str, force: bool = False) -> None:
         """Go through the post processing steps for a flowcell