+++ cg/models/demultiplex/demux_results.py
@@ -64,7 +64,18 @@ class DemuxResults:
 
     @property
     def sample_sheet_path(self) -> Path:
+        """Return the path to where the original sample sheet is"""
         return self.flowcell.sample_sheet_path
+
+    @property
+    def demux_sample_sheet_path(self) -> Path:
+        """Return the path to sample sheet in demuxed flowcell dir"""
+        return self.results_dir / self.flowcell.sample_sheet_path.name
+
+    @property
+    def copy_complete_path(self) -> Path:
+        """Return the path to a file named copycomplete.txt used as flag that post processing is ready"""
+        return self.demux_dir / "copycomplete.txt"
 
     @property
     def projects(self) -> Iterable[str]: